Shared steps class is removed
