+++ src/test/java/com/azulCRM/pages/LoginPage.java
@@ -2,10 +2,8 @@ package com.azulCRM.pages;
 
 
 import com.azulCRM.utilities.ConfigurationReader;
-import com.azulCRM.utilities.Driver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
-import org.openqa.selenium.support.PageFactory;
 
 public class LoginPage extends BasePage {
 
@@ -20,9 +18,9 @@ public class LoginPage extends BasePage {
 
     public void login(String userType) {
         String username = ConfigurationReader.getProperty(userType + "_username");
-        String user_password = ConfigurationReader.getProperty("password");
-        userName.sendKeys(username);
-        password.sendKeys(user_password);
+        String password = ConfigurationReader.getProperty("password");
+        this.userName.sendKeys(username);
+        this.password.sendKeys(password);
         submit.click();
     }
 }
